requirements.txt
streamlit>=1.28.0
plotly>=5.14.0
numpy>=1.21.0
Pillow>=9.0.0
